requirements.txt
APScheduler==3.8.1
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
pyTelegramBotAPI==4.3.1
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
